cleanup codegen conftest
