+++ .github/workflows/test-graphql-codegen.yaml
@@ -3,14 +3,18 @@ name: Test GraphQL code generation
 on:
   push:
     paths:
-      - '.github/workflows/test-quilt3-admin-codegen.yaml'
+      - '.github/workflows/test-graphql-codegen.yaml'
       - 'shared/graphql/schema.graphql'
+      - 'api/python/quilt3-admin/**'
+      - 'api/python/quilt3/admin/_graphql_client/**'
       - 'api/python/quilt3-graphql/**'
       - 'api/python/quilt3/_graphql_client/**'
   pull_request:
     paths:
-      - '.github/workflows/test-quilt3-admin-codegen.yaml'
+      - '.github/workflows/test-graphql-codegen.yaml'
       - 'shared/graphql/schema.graphql'
+      - 'api/python/quilt3-admin/**'
+      - 'api/python/quilt3/admin/_graphql_client/**'
       - 'api/python/quilt3-graphql/**'
       - 'api/python/quilt3/_graphql_client/**'
   merge_group:
